Inlining simplified byte-addition
This is Dave's idea for streamlining, with additional safety check.

We're adding one byte to a word, so there could be a carry of a single unit up through several bytes.

If the sign of a given byte is unchanged, then it cannot have carried: in order to carry, it would have to have been FF+1=100 or 00-1=1FF

It turns out the 4 byte word is good enough for thousands of digits of pi.  An assembly language implementation would chase any carries up as far as is needed - in Basic that's messy.